Rename lib by types
